pyproject.toml
[tool.poetry]
name = "bd-test"
version = "0.1.0"
description = ""
authors = ["Nikita <erofeev.nik.it@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = {extras = ["all"], version = "^2.0.23"}
black = "^23.11.0"
envparse = "^0.2.0"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.9"
alembic = "^1.13.0"
uvicorn = "^0.24.0.post1"
fastapi = "^0.104.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
